+++ src/System/Database/MyQuery/Table.php
@@ -62,7 +62,7 @@ class Table
         TABLE_SCHEMA = :dbs AND TABLE_NAME = :table"
     );
     $this->PDO->bind(':table', $this->table_name);
-    $this->PDO->bind(':dbs', DB_NAME);
+    $this->PDO->bind(':dbs', $this->PDO->configs()['database_name']);
     return $this->PDO->resultset() ?? array();
   }
 }