+++ CHANGELOG.md
@@ -5,9 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- Added string manipulation `Str::chartAt`, `Str::concat`, `Str::indexOf`, `Str::lastIndexOf`, `Str::match`, `Str::repalce`, `Str::search`, `Str::slice`, `Str::split`, `Str::toLowerCase`, `Str::toUpperCase`, `Str::firstUpper`, `Str::firstUppeAll`, `Str::toPascalCase`, `Str::toCamelCase`, `Str::contains`, `Str::startsWith`, `Str::endsWith`, `Str::template`, `Str::slug`, `Str::repeat`, `Str::length`, `Str::isString`, `Str::isEmpty`.
+- Added string macro `Str::macro`.
+
+## [0.6.2] - 2022-05-27
 ### Changed
 - New `MyPDO::__construct` params, using array contain database configuration.
 - `MyPDO::getInstance()` return single instance.
+- New `MyPDO::__construct` throwing error without call `die` function.
 
 ### Removed
 - Remove database defineder (`DB_NAME`, `DB_HOST`, `DB_USER`, `DB_PASS`).